Refactor code for clarity and to separate the fiber object and add some business logic.
